+++ angry/src/main/java/br/com/angryapps/angry/models/Card.java
@@ -14,7 +14,7 @@ public class Card {
     private UUID id;
     private String title;
     private String description;
-    private int position;
+    private double position;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
@@ -50,11 +50,11 @@ public class Card {
         this.description = description;
     }
 
-    public int getPosition() {
+    public double getPosition() {
         return position;
     }
 
-    public void setPosition(int position) {
+    public void setPosition(double position) {
         this.position = position;
     }
